refactor(cli): extract playwright redaction to playwright_redaction.py (#1331)
Move the subprocess-output redaction cluster (the _ANSI_* regexes,
_REDACTION_SAFE_VALUES, _REDACTION_MIN_VALUE_LEN, _strip_ansi,
_expand_nested_secret_values, redact_subprocess_output) plus its leading
section comment out of cli/services/playwright_login.py into a new
cli/services/playwright_redaction.py. playwright_login.py re-imports
redact_subprocess_output near the top so __all__ and any
mock.patch("...playwright_login.redact_subprocess_output") keep working;
run_playwright_login still calls it via that module-level name.

Drops playwright_login.py from 988 to 842 LOC (off the size ratchet
allowlist); new module is 150 LOC. Pure relocation, no behavior change.

Also: add playwright_redaction.py to the CLAUDE.md Repository Structure
tree and to test_services_boundary.py GUARDED_PATHS (pure string funcs,
Click-free).
