+++ tests/_guardrails/test_module_size_ratchet.py
@@ -68,7 +68,6 @@ ALLOWLISTED_CEILINGS: dict[str, int] = {
     "_source/upload.py": 1236,
     "cli/session_cmd.py": 1080,
     "_sources.py": 1007,
-    "cli/services/playwright_login.py": 988,
     "_artifact/downloads.py": 973,
     "client.py": 973,
     "_research.py": 936,
